Refactor layout for content summarization
