+++ src/main/java/nl/infosupport/agenttalks/reader/ContentSummarizer.java
@@ -1,9 +1,8 @@
-package nl.infosupport.agenttalks.reader.service;
+package nl.infosupport.agenttalks.reader;
 
 import dev.langchain4j.service.SystemMessage;
 import dev.langchain4j.service.UserMessage;
 import io.quarkiverse.langchain4j.RegisterAiService;
-import nl.infosupport.agenttalks.reader.model.ContentSummary;
 
 @RegisterAiService
 public interface ContentSummarizer {
