+++ src/main/java/nl/infosupport/agenttalks/reader/ContentSummary.java
@@ -1,4 +1,4 @@
-package nl.infosupport.agenttalks.reader.model;
+package nl.infosupport.agenttalks.reader;
 
 public class ContentSummary {
     public String title;
